# packages/core/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kinora API",
    "summary": "Generation-on-scroll showrunner backend.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/books": {
      "get": {
        "tags": [
          "books"
        ],
        "summary": "List Books",
        "description": "List the books the current user owns (the shelf), newest first.",
        "operationId": "list_books_api_books_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BookResponse"
                  },
                  "type": "array",
                  "title": "Response List Books Api Books Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "books"
        ],
        "summary": "Upload Book",
        "description": "Validate + store a PDF/EPUB, create the book, and trigger Phase A ingest.\n\nReturns the freshly-created book (status ``importing``) directly \u2014 the shelf\nprepends it and polls until it is ``ready``.\n\nDefends the ingest pipeline: the content-type is checked, the per-user book\nquota is enforced, the body is read **streaming** (aborting the moment it\ncrosses ``MAX_PDF_BYTES`` rather than buffering an oversized payload), the\nformat is detected by content magic, an EPUB is normalised to PDF, and the\npage count is capped (``MAX_INGEST_PAGES``) before any token-spending ingest\nis triggered.",
        "operationId": "upload_book_api_books_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_book_api_books_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/books/{book_id}": {
      "get": {
        "tags": [
          "books"
        ],
        "summary": "Get Book",
        "description": "Fetch a book with its import status + progress.",
        "operationId": "get_book_api_books__book_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Book Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/books/{book_id}/pages/{page_number}": {
      "get": {
        "tags": [
          "books"
        ],
        "summary": "Get Page",
        "description": "A page's presigned image URL, text, and per-word boxes (\u00a79.4).",
        "operationId": "get_page_api_books__book_id__pages__page_number__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Book Id"
            }
          },
          {
            "name": "page_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Page Number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/books/{book_id}/canon": {
      "get": {
        "tags": [
          "books"
        ],
        "summary": "Get Canon",
        "description": "The book's canon graph: the entity list the Director editor renders, plus\nthe human-inspectable markdown vault export (\u00a78.1).",
        "operationId": "get_canon_api_books__book_id__canon_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Book Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CanonResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/books/{book_id}/shots": {
      "get": {
        "tags": [
          "books"
        ],
        "summary": "List Shots",
        "description": "The book's shots (the \u00a75.4 shot timeline) as a bare array.",
        "operationId": "list_shots_api_books__book_id__shots_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Book Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ShotResponse"
                  },
                  "title": "Response List Shots Api Books  Book Id  Shots Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_upload_book_api_books_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "The source PDF or EPUB"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "author": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author"
          },
          "art_direction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Art Direction"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_book_api_books_post"
      },
      "BookResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "author": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "num_pages": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Num Pages"
          },
          "art_direction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Art Direction"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "progress": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress"
          },
          "stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "status"
        ],
        "title": "BookResponse",
        "description": "A book on the shelf, with its import status and progress (\u00a75.1).\n\nThe collection endpoints (``GET /books``) return a *bare* ``list`` of these,\nand the upload endpoint returns a bare ``BookResponse`` \u2014 the frontend client\nparses ``Book[]`` / ``Book`` directly (no envelope)."
      },
...
      "CanonResponse": {
        "properties": {
          "book_id": {
            "type": "string",
            "title": "Book Id"
          },
          "entities": {
            "items": {
              "$ref": "#/components/schemas/CanonEntityResponse"
            },
            "type": "array",
            "title": "Entities"
          },
          "states": {
            "items": {
              "$ref": "#/components/schemas/CanonStateResponse"
            },
            "type": "array",
            "title": "States"
          },
          "markdown": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Markdown"
          }
        },
        "type": "object",
        "required": [
          "book_id"
        ],
        "title": "CanonResponse",
        "description": "The canon graph for a book: the entity list the Director editor renders,\nthe versioned continuity facts (\u00a78.5), plus the optional human-inspectable\nmarkdown vault export (\u00a78.1)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PageResponse": {
        "properties": {
          "book_id": {
            "type": "string",
            "title": "Book Id"
          },
          "page_number": {
            "type": "integer",
            "title": "Page Number"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "word_boxes": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Word Boxes"
          }
        },
        "type": "object",
        "required": [
          "book_id",
          "page_number"
        ],
        "title": "PageResponse",
        "description": "One rendered page: a presigned image URL, text, and per-word boxes (\u00a79.4)."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}